Add AssignRequest to PreImage.IsAvailableFor
While Microsoft has announced that they are deprecating the AssignRequest, it is still in active use, even by MS themselves.

The Assign button inside the Ribbon, available on specific entities in model-driven apps, does not use the Update message - as would be preferred - but rather uses the Assign message.

The Assign message in turn does support PreImages when used for Plugin Registration. In order to properly reflect this I suggest adding it to the "IsAvailableFor" method in the PreImage class.